extensions/quic/requirements.txt
aioquic
pyjwt
requests
cryptography